pyproject.toml
[project]
name = "agno-template"
version = "0.1.0"
description = "A template project using Agno"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno==2.0.3",
    "annotated-types==0.7.0",
    "anthropic==0.67.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "email-validator==2.3.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.11",
    "fastapi-cloud-cli==0.1.5",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-auth==2.40.3",
    "google-genai==1.36.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mcp==1.13.1",
    "mdurl==0.1.2",
    "mem0ai>=0.1.116",
    "openai==1.107.1",
    "packaging==25.0",
    "pgvector>=0.4.1",
    "psycopg[binary]==3.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rich==14.1.0",
    "rich-toolkit==0.15.1",
    "rignore==0.6.4",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "sentry-sdk==2.37.1",
    "shellingham==1.5.4",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "starlette==0.47.3",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typer==0.17.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]

[tool.uv]
dev-dependencies = []
